PR B: extract DayCalendar into pure DayCalendarView + thin authed container (sheds useDayFeed + entity context)
